# artifacts/api-server/src/routes/devis.ts
import { Router, type IRouter } from "express";
import { eq, sql } from "drizzle-orm";
import { db, devisTable, lignesDevisTable, atelierTable, facturesTable, lignesFactureTable } from "@workspace/db";
import { execRows, serializeDates } from "../lib/db-utils";
import {
  ListDevisResponse,
  CreateDevisBody,
  GetDevisParams,
  GetDevisResponse,
  DeleteDevisParams,
  DeleteDevisResponse,
  UpdateDevisStatutParams,
  UpdateDevisStatutBody,
  UpdateDevisStatutResponse,
  SaveDevisLignesParams,
  SaveDevisLignesBody,
  SaveDevisLignesResponse,
  ConvertDevisToFactureParams,
} from "@workspace/api-zod";

const router: IRouter = Router();

const parseNum = (v: unknown) => parseFloat(String(v ?? "0"));

// ─── Helper: calculate line quantity ──────────────────────────────────────────
function calcLigne(unite: string, largeur: number | null, hauteur: number | null, quantite: number) {
  if (unite === "metre_lineaire") {
    return ((largeur ?? 0) + (hauteur ?? 0)) * 2 * quantite;
  }
  if (unite === "metre_carre") {
    return (largeur ?? 0) * (hauteur ?? 0) * quantite;
  }
  return quantite;
}

// ─── Helper: get next document number ─────────────────────────────────────────
async function getNextNumero(type: "devis" | "facture"): Promise<string> {
  const [atelier] = await db.select().from(atelierTable).where(eq(atelierTable.id, 1));
  if (!atelier) throw new Error("Atelier non configuré");

  const year = new Date().getFullYear();

  if (type === "devis") {
    const next = (atelier.compteur_devis ?? 0) + 1;
    await db.update(atelierTable).set({ compteur_devis: next }).where(eq(atelierTable.id, 1));
    return `${atelier.prefixe_devis}-${year}-${String(next).padStart(3, "0")}`;
  } else {
    const next = (atelier.compteur_facture ?? 0) + 1;
    await db.update(atelierTable).set({ compteur_facture: next }).where(eq(atelierTable.id, 1));
    return `${atelier.prefixe_facture}-${year}-${String(next).padStart(3, "0")}`;
  }
}

// ─── Helper: recalculate devis totals ─────────────────────────────────────────
async function recalcDevis(devisId: number): Promise<void> {
  const lignes = await db.select().from(lignesDevisTable).where(eq(lignesDevisTable.devis_id, devisId));

  let ht = 0, tva10 = 0, tva20 = 0;
  for (const l of lignes) {
    ht += l.total_ht;
    if (l.taux_tva === 10) tva10 += l.total_ht * 0.1;
    else tva20 += l.total_ht * 0.2;
  }

  await db.update(devisTable)
    .set({ sous_total_ht: ht, total_tva_10: tva10, total_tva_20: tva20, total_ttc: ht + tva10 + tva20, modifie_le: new Date() })
    .where(eq(devisTable.id, devisId));
}

// ─── Helper: get devis with client name ───────────────────────────────────────
type DevisWithClient = {
  id: number; numero: string; client_id: number; client_nom: string; client_prenom: string;
  date_creation: string; date_validite: string; statut: string; sous_total_ht: string;
  total_tva_10: string; total_tva_20: string; total_ttc: string; notes: string;
  conditions: string; facture_id: number; cree_le: string; modifie_le: string;
};

async function getDevisWithClient(id: number): Promise<DevisWithClient | undefined> {
  const rows = await execRows<DevisWithClient>(
    sql`SELECT d.*, c.nom as client_nom, c.prenom as client_prenom
        FROM devis d LEFT JOIN clients c ON c.id = d.client_id WHERE d.id = ${id}`
  );
  return rows[0];
}

function mapDevis(r: DevisWithClient) {
  const s = serializeDates(r as unknown as Record<string, unknown>);
  return {
    id: r.id, numero: r.numero, client_id: r.client_id, client_nom: r.client_nom ?? null,
    client_prenom: r.client_prenom ?? null, date_creation: r.date_creation,
    date_validite: r.date_validite ?? null, statut: r.statut,
    sous_total_ht: parseNum(r.sous_total_ht), total_tva_10: parseNum(r.total_tva_10),
    total_tva_20: parseNum(r.total_tva_20), total_ttc: parseNum(r.total_ttc),
    notes: r.notes ?? null, conditions: r.conditions ?? null, facture_id: r.facture_id ?? null,
    cree_le: s.cree_le as string, modifie_le: s.modifie_le as string,
  };
}

// ─── Routes ───────────────────────────────────────────────────────────────────

router.get("/devis", async (req, res): Promise<void> => {
  const statut = typeof req.query.statut === "string" ? req.query.statut : null;
  const clientId = typeof req.query.client_id === "string" ? parseInt(req.query.client_id, 10) : null;

  let query = sql`SELECT d.*, c.nom as client_nom, c.prenom as client_prenom
      FROM devis d LEFT JOIN clients c ON c.id = d.client_id WHERE 1=1`;

  if (statut) query = sql`${query} AND d.statut = ${statut}`;
  if (clientId) query = sql`${query} AND d.client_id = ${clientId}`;
  query = sql`${query} ORDER BY d.cree_le DESC`;

  const rows = await execRows<DevisWithClient>(query);
  res.json(ListDevisResponse.parse(rows.map(mapDevis)));
});

router.post("/devis", async (req, res): Promise<void> => {
  const parsed = CreateDevisBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  // Ensure atelier exists
  const [existingAtelier] = await db.select().from(atelierTable).where(eq(atelierTable.id, 1));
  if (!existingAtelier) {
    await db.insert(atelierTable).values({ id: 1, nom: "Mon Atelier" });
  }

  const numero = await getNextNumero("devis");
  const validite = new Date();
  validite.setDate(validite.getDate() + 30);

  const [devis] = await db.insert(devisTable).values({
    numero,
    client_id: parsed.data.client_id,
    notes: parsed.data.notes ?? null,
    statut: "brouillon",
    date_validite: validite.toISOString().slice(0, 10),
  }).returning();

  const withClient = await getDevisWithClient(devis.id);
  res.status(201).json(SaveDevisLignesResponse.parse(mapDevis(withClient!)));
});

router.get("/devis/:id", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = GetDevisParams.safeParse({ id: parseInt(raw, 10) });
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }

  const devisRow = await getDevisWithClient(params.data.id);
  if (!devisRow) {
    res.status(404).json({ error: "Devis introuvable" });
    return;
  }

  const lignes = await db.select().from(lignesDevisTable)
    .where(eq(lignesDevisTable.devis_id, params.data.id))
    .orderBy(lignesDevisTable.ordre);

  const data = {
    ...mapDevis(devisRow),
    lignes: lignes.map((l) => ({
      id: l.id, devis_id: l.devis_id, produit_id: l.produit_id ?? null,
      designation: l.designation, unite_calcul: l.unite_calcul,
      largeur_m: l.largeur_m ?? null, hauteur_m: l.hauteur_m ?? null,
      quantite: l.quantite, quantite_calculee: l.quantite_calculee ?? null,
      prix_unitaire_ht: l.prix_unitaire_ht, taux_tva: l.taux_tva,
      total_ht: l.total_ht, total_ttc: l.total_ttc, ordre: l.ordre,
    })),
  };

  res.json(GetDevisResponse.parse(data));
});

router.delete("/devis/:id", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = DeleteDevisParams.safeParse({ id: parseInt(raw, 10) });
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }

  await db.delete(devisTable).where(eq(devisTable.id, params.data.id));
  res.json(DeleteDevisResponse.parse({ success: true }));
});

router.patch("/devis/:id/statut", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = UpdateDevisStatutParams.safeParse({ id: parseInt(raw, 10) });
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }

  const parsed = UpdateDevisStatutBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  await db.update(devisTable)
    .set({ statut: parsed.data.statut, modifie_le: new Date() })
    .where(eq(devisTable.id, params.data.id));

  const updated = await getDevisWithClient(params.data.id);
  if (!updated) {
    res.status(404).json({ error: "Devis introuvable" });
    return;
  }

  res.json(UpdateDevisStatutResponse.parse(mapDevis(updated)));
});

router.put("/devis/:id/lignes", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = SaveDevisLignesParams.safeParse({ id: parseInt(raw, 10) });
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }

  const parsed = SaveDevisLignesBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const devisId = params.data.id;

  // Delete all existing lines then re-insert
  await db.delete(lignesDevisTable).where(eq(lignesDevisTable.devis_id, devisId));

  for (let i = 0; i < parsed.data.lignes.length; i++) {
    const l = parsed.data.lignes[i];
    const qCalc = calcLigne(l.unite_calcul, l.largeur_m ?? null, l.hauteur_m ?? null, l.quantite);
    const totalHT = qCalc * l.prix_unitaire_ht;
    const totalTTC = totalHT * (1 + l.taux_tva / 100);

    await db.insert(lignesDevisTable).values({
      devis_id: devisId,
      produit_id: l.produit_id ?? null,
      designation: l.designation,
      unite_calcul: l.unite_calcul,
      largeur_m: l.largeur_m ?? null,
      hauteur_m: l.hauteur_m ?? null,
      quantite: l.quantite,
      quantite_calculee: qCalc,
      prix_unitaire_ht: l.prix_unitaire_ht,
      taux_tva: l.taux_tva,
      total_ht: totalHT,
      total_ttc: totalTTC,
      ordre: l.ordre ?? i,
    });
  }

  await recalcDevis(devisId);

  const updated = await getDevisWithClient(devisId);
  if (!updated) {
    res.status(404).json({ error: "Devis introuvable" });
    return;
  }

  res.json(SaveDevisLignesResponse.parse(mapDevis(updated)));
});

router.post("/devis/:id/convertir", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = ConvertDevisToFactureParams.safeParse({ id: parseInt(raw, 10) });
  if (!params.success) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }

  const [devis] = await db.select().from(devisTable).where(eq(devisTable.id, params.data.id));
  if (!devis) {
    res.status(404).json({ error: "Devis introuvable" });
    return;
  }

  const numero = await getNextNumero("facture");
  const echeance = new Date();
  echeance.setDate(echeance.getDate() + 30);

  const [facture] = await db.insert(facturesTable).values({
    numero,
    devis_id: devis.id,
    client_id: devis.client_id,
    sous_total_ht: devis.sous_total_ht,
    total_tva_10: devis.total_tva_10,
    total_tva_20: devis.total_tva_20,
    total_ttc: devis.total_ttc,
    total_paye: 0,
    solde_restant: devis.total_ttc,
    notes: devis.notes,
    conditions: devis.conditions,
    statut: "brouillon",
    date_echeance: echeance.toISOString().slice(0, 10),
  }).returning();

  // Copy lines from devis to facture
  const lignes = await db.select().from(lignesDevisTable).where(eq(lignesDevisTable.devis_id, devis.id));
  for (const l of lignes) {
    await db.insert(lignesFactureTable).values({
      facture_id: facture.id,
      produit_id: l.produit_id,
      designation: l.designation,
      unite_calcul: l.unite_calcul,
      largeur_m: l.largeur_m,
      hauteur_m: l.hauteur_m,
      quantite: l.quantite,
      quantite_calculee: l.quantite_calculee,
      prix_unitaire_ht: l.prix_unitaire_ht,
      taux_tva: l.taux_tva,
      total_ht: l.total_ht,
      total_ttc: l.total_ttc,
      ordre: l.ordre,
    });
  }

  // Mark devis as converted
  await db.update(devisTable)
    .set({ statut: "converti", facture_id: facture.id, modifie_le: new Date() })
    .where(eq(devisTable.id, devis.id));

  const factureRows = await execRows<{
    id: number; numero: string; devis_id: number; client_id: number;
    client_nom: string; client_prenom: string; date_creation: string; date_echeance: string;
    statut: string; sous_total_ht: string; total_tva_10: string; total_tva_20: string;
    total_ttc: string; total_paye: string; solde_restant: string; notes: string;
    conditions: string; cree_le: string; modifie_le: string;
  }>(
    sql`SELECT f.*, c.nom as client_nom, c.prenom as client_prenom
        FROM factures f LEFT JOIN clients c ON c.id = f.client_id WHERE f.id = ${facture.id}`
  );

  const f = factureRows[0];
  res.status(201).json({
    id: f.id, numero: f.numero, devis_id: f.devis_id ?? null,
    client_id: f.client_id, client_nom: f.client_nom ?? null,
    client_prenom: f.client_prenom ?? null, date_creation: f.date_creation,
    date_echeance: f.date_echeance ?? null, statut: f.statut,
    sous_total_ht: parseNum(f.sous_total_ht), total_tva_10: parseNum(f.total_tva_10),
    total_tva_20: parseNum(f.total_tva_20), total_ttc: parseNum(f.total_ttc),
    total_paye: parseNum(f.total_paye), solde_restant: parseNum(f.solde_restant),
    notes: f.notes ?? null, conditions: f.conditions ?? null,
    cree_le: f.cree_le, modifie_le: f.modifie_le,
  });
});

export default router;