# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: EncadrePro API – gestion d'un atelier d'encadrement
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devis:
    get:
      operationId: listDevis
      tags: [devis]
      summary: List quotes
      parameters:
        - name: statut
          in: query
          schema:
            type: string
        - name: client_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Devis"

    post:
      operationId: createDevis
      tags: [devis]
      summary: Create a quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDevisBody"
      responses:
        "201":
          description: Created quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Devis"
  /devis/{id}:
    get:
      operationId: getDevis
      tags: [devis]
      summary: Get a quote with its lines
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote with lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevisDetail"

    delete:
      operationId: deleteDevis
      tags: [devis]
      summary: Delete a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /devis/{id}/statut:
    patch:
      operationId: updateDevisStatut
      tags: [devis]
      summary: Update quote status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatutBody"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Devis"
  /devis/{id}/lignes:
    put:
      operationId: saveDevisLignes
      tags: [devis]
      summary: Save all lines for a quote (replaces existing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveLignesBody"
      responses:
        "200":
          description: Updated quote with recalculated totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Devis"
  /devis/{id}/convertir:
    post:
      operationId: convertDevisToFacture
      tags: [devis]
      summary: Convert quote to invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Created invoice from quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facture"

  # ──────────────── FACTURES ────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── DASHBOARD ──
...
    LigneDevis:
      type: object
      properties:
        id: { type: integer }
        devis_id: { type: integer }
        produit_id: { type: integer, nullable: true }
        designation: { type: string }
        unite_calcul: { type: string }
        largeur_m: { type: number, nullable: true }
        hauteur_m: { type: number, nullable: true }
        quantite: { type: number }
        quantite_calculee: { type: number, nullable: true }
        prix_unitaire_ht: { type: number }
        taux_tva: { type: number }
        total_ht: { type: number }
        total_ttc: { type: number }
        ordre: { type: integer }
      required: [id, devis_id, designation, unite_calcul, quantite, prix_unitaire_ht, taux_tva, total_ht, total_ttc, ordre]
    LigneInput:
      type: object
      properties:
        produit_id: { type: integer, nullable: true }
        designation: { type: string }
        unite_calcul: { type: string }
        largeur_m: { type: number, nullable: true }
        hauteur_m: { type: number, nullable: true }
        quantite: { type: number }
        prix_unitaire_ht: { type: number }
        taux_tva: { type: number }
        ordre: { type: integer }
      required: [designation, unite_calcul, quantite, prix_unitaire_ht, taux_tva]
    Devis:
      type: object
      properties:
        id: { type: integer }
        numero: { type: string }
        client_id: { type: integer }
        client_nom: { type: string, nullable: true }
        client_prenom: { type: string, nullable: true }
        date_creation: { type: string }
        date_validite: { type: string, nullable: true }
        statut: { type: string }
        sous_total_ht: { type: number }
        total_tva_10: { type: number }
        total_tva_20: { type: number }
        total_ttc: { type: number }
        notes: { type: string, nullable: true }
        conditions: { type: string, nullable: true }
        facture_id: { type: integer, nullable: true }
        cree_le: { type: string }
        modifie_le: { type: string }
      required: [id, numero, client_id, date_creation, statut, sous_total_ht, total_tva_10, total_tva_20, total_ttc, cree_le, modifie_le]
    DevisDetail:
      allOf:
        - $ref: "#/components/schemas/Devis"
        - type: object
          properties:
            lignes:
              type: array
              items:
                $ref: "#/components/schemas/LigneDevis"
          required: [lignes]
    CreateDevisBody:
      type: object
      properties:
        client_id: { type: integer }
        notes: { type: string, nullable: true }
      required: [client_id]
    SaveLignesBody:
      type: object
      properties:
        lignes:
          type: array
          items:
            $ref: "#/components/schemas/LigneInput"
      required: [lignes]
    UpdateStatutBody:
      type: object
      properties:
        statut: { type: string }
      required: [statut]

    # ── FACTURES ──
...
    Facture:
      type: object
      properties:
        id: { type: integer }
        numero: { type: string }
        devis_id: { type: integer, nullable: true }
        client_id: { type: integer }
        client_nom: { type: string, nullable: true }
        client_prenom: { type: string, nullable: true }
        date_creation: { type: string }
        date_echeance: { type: string, nullable: true }
        statut: { type: string }
        sous_total_ht: { type: number }
        total_tva_10: { type: number }
        total_tva_20: { type: number }
        total_ttc: { type: number }
        total_paye: { type: number }
        solde_restant: { type: number }
        notes: { type: string, nullable: true }
        conditions: { type: string, nullable: true }
        cree_le: { type: string }
        modifie_le: { type: string }
      required: [id, numero, client_id, date_creation, statut, sous_total_ht, total_tva_10, total_tva_20, total_ttc, total_paye, solde_restant, cree_le, modifie_le]
...