# lib/api-zod/src/generated/types/updateStatutBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EncadrePro API – gestion d'un atelier d'encadrement
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateStatutBody {
  statut: string;
}